+++ src/main/web/components/forms/inputs/dropdown/DropdownWithFilter.scss
@@ -1,0 +1,61 @@
+/**
+ * ResearchSpace
+ * Copyright (C) 2022-2024, © Kartography Community Interest Company
+ * Copyright (C) 2020, © Trustees of the British Museum
+ * Copyright (C) 2015-2019, metaphacts GmbH
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+.dropdown-with-filter {
+  display: inline-block;
+  position: relative;
+  margin-left: 8px;
+}
+
+.dropdown-with-filter .dropdown-toggle {
+  user-select: none;
+}
+
+.dropdown-with-filter .dropdown-menu {
+  min-width: 180px;
+  max-height: 240px;
+  overflow-y: auto;
+  padding: 0;
+}
+
+.dropdown-with-filter .dropdown-filter-input {
+  padding: 8px 12px;
+  position: sticky;
+  top: 0;
+  background: #fff;
+  z-index: 1;
+}
+
+.dropdown-with-filter .dropdown-filter-input input {
+  width: 100%;
+  margin-bottom: 4px;
+}
+
+.dropdown-with-filter .dropdown-item {
+  display: block;
+  padding: 6px 20px;
+  cursor: pointer;
+  white-space: normal;
+}
+
+.dropdown-with-filter .dropdown-no-results {
+  color: #888;
+  padding: 6px 20px;
+  display: block;
+}
